+++ test/DebugInfo/PrivateDiscriminator.swift
@@ -9,14 +9,14 @@
 func markUsed<T>(t: T) {}
 
 private class A {
-  init(val : Int) { member = val }
-  private let member : Int
+  init(val : Int64) { member = val }
+  private let member : Int64
   // CHECK: !DISubprogram(name: "getMember"
   // CHECK-SAME:          linkageName: "{{[^"]*}}[[DISCRIMINATOR]]
   // CHECK-SAME:          line: [[@LINE+2]]
   // CHECK-SAME:          isLocal: true, isDefinition: true 
-  private func getMember() -> Int { return member }
-  func getVal() -> Int { return getMember() }  
+  private func getMember() -> Int64 { return member }
+  func getVal() -> Int64 { return getMember() }  
 }
 
 func f() {
