+++ test/DebugInfo/fnptr.swift
@@ -1,14 +1,14 @@
 // RUN: %target-swift-frontend %s -emit-ir -g -o - | FileCheck %s
 
-// CHECK-DAG: ![[SINODE:.*]] = !DICompositeType(tag: DW_TAG_structure_type, name: "Int",{{.*}} identifier: [[SI:.*]])
+// CHECK-DAG: ![[SINODE:.*]] = !DICompositeType(tag: DW_TAG_structure_type, name: "Int64",{{.*}} identifier: [[SI:.*]])
 // CHECK-DAG: ![[SFNODE:.*]] = !DICompositeType(tag: DW_TAG_structure_type, name: "Float",{{.*}} identifier: [[SF:.*]])
 // CHECK-DAG: ![[VOIDNODE:.*]] = !DICompositeType(tag: DW_TAG_structure_type, name: "_TtT_",{{.*}} identifier: [[VOID:.*]])
 func bar() {
     print("bar()", terminator: "")
 }
-func baz(i: Float) -> Int { return 0; }
-func barz(i: Float, _ j: Float) -> Int { return 0; }
-func main() -> Int {
+func baz(i: Float) -> Int64 { return 0; }
+func barz(i: Float, _ j: Float) -> Int64 { return 0; }
+func main() -> Int64 {
 
     // CHECK-DAG: !DILocalVariable(tag: DW_TAG_auto_variable, name: "bar_function_pointer",{{.*}} line: [[@LINE+1]],{{.*}} type: !"[[BARPT:[^,]+]]"
     var bar_function_pointer = bar
@@ -26,7 +26,7 @@ func main() -> Int {
     // CHECK-DAG: ![[BAZMEMBER]] = !DIDerivedType(tag: DW_TAG_member,{{.*}} baseType: ![[BAZPTR:[0-9]+]]
     // CHECK-DAG: ![[BAZPTR]] = !DIDerivedType(tag: DW_TAG_pointer_type,{{.*}} baseType: ![[BAZT:[0-9]+]]
     // CHECK-DAG: ![[BAZT]] = !DISubroutineType(types: ![[BAZARGS:.*]])
-    // CHECK-DAG: ![[BAZARGS]] = !{!"_TtSi", !"_TtSf"}
+    // CHECK-DAG: ![[BAZARGS]] = !{!"_TtVSs5Int64", !"_TtSf"}
     var baz_function_pointer = baz
     baz_function_pointer(2.89)
 
@@ -36,7 +36,7 @@ func main() -> Int {
     // CHECK-DAG: ![[BARZMEMBER]] = !DIDerivedType(tag: DW_TAG_member,{{.*}} baseType: ![[BARZPTR:[0-9]+]]
     // CHECK-DAG: ![[BARZPTR]] = !DIDerivedType(tag: DW_TAG_pointer_type,{{.*}} baseType: ![[BARZT:[0-9]+]]
     // CHECK-DAG: ![[BARZT]] = !DISubroutineType(types: ![[BARZARGS:.*]])
-    // CHECK-DAG: ![[BARZARGS]] = !{!"_TtSi", !"_TtSf", !"_TtSf"}
+    // CHECK-DAG: ![[BARZARGS]] = !{!"_TtVSs5Int64", !"_TtSf", !"_TtSf"}
     var barz_function_pointer = barz
     return barz_function_pointer(2.89, -1.0)
 }
