+++ test/DebugInfo/nested_functions.swift
@@ -1,12 +1,12 @@
 // RUN: %target-swift-frontend -primary-file %s -emit-ir -g -o - | FileCheck %s
 
-func outer(a: Int) -> Int {
+func outer(a: Int64) -> Int64 {
      // Inner functions have a linkage name of "closure[0-9]+", but
      // their DW_AT_name is preserved.
 
-     // CHECK: !DISubprogram(name: "inner", linkageName: "_TFF16nested_functions5outerFSiSiL_5innerfSiSi"
+     // CHECK: !DISubprogram(name: "inner", linkageName: "_TFF16nested_functions5outerFVSs5Int64S0_L_5innerfS0_S0_"
      // CHECK-SAME:          line: [[@LINE+1]]
-     func inner(b: Int) -> Int {
+     func inner(b: Int64) -> Int64 {
        return a+b
      }
