+++ test/DebugInfo/pcomp.swift
@@ -17,15 +17,15 @@ func f(arg :protocol<A,B>) {
 
 
 protocol SomeProto {
-  func f() -> Int
+  func f() -> Int64
 }
 
 class SomeClass : SomeProto {
-  func f() -> Int { return 1 }
+  func f() -> Int64 { return 1 }
 }
 
 class SomeOtherClass : SomeClass {
-  override func f() -> Int { return 1 }
+  override func f() -> Int64 { return 1 }
 }
 // This is an indirect value.
 // CHECK-DAG: !DICompositeType(tag: DW_TAG_structure_type, name: "SomeProto",{{.*}} identifier: "_TtP5pcomp9SomeProto_"
